Effect Name property is now the same as in SoX command line
Also changed the way effect Name is stored. Now each BaseEffect
specialization shall define the Name  property to return the name
instead of initializing the Name property of BaseEffect. This allow to
define it only once on each class. With the previous method, if a
specialization has multiple constructors, there is a high chance of
needing to call base(“name”) several times.